tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.fillStyle.relativecolor.currentcolor.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.fillStyle.relativecolor.currentcolor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.fillStyle.relativecolor.currentcolor</h1>
<p class="desc">Relative color works as color input with currentcolor</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Relative color works as color input with currentcolor");
_addTest(function(canvas, ctx) {

  canvas.setAttribute('style', 'color: magenta');
  ctx.fillStyle = "rgb(from currentcolor g r b)";
  _assertSame(ctx.fillStyle, 'color(srgb 0 1 1)', "ctx.fillStyle", "'color(srgb 0 1 1)'");
  ctx.strokeStyle = "rgb(from currentcolor g r b)";
  _assertSame(ctx.strokeStyle, 'color(srgb 0 1 1)', "ctx.strokeStyle", "'color(srgb 0 1 1)'");

});
</script>
